+++ ehpc-20240730/src/Models/ListJobsRequest/jobFilter/diagnosis.php
@@ -1,0 +1,71 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\EHPC\V20240730\Models\ListJobsRequest\jobFilter;
+
+use AlibabaCloud\Tea\Model;
+
+class diagnosis extends Model
+{
+    /**
+     * @var string
+     */
+    public $operator;
+
+    /**
+     * @var string
+     */
+    public $option;
+
+    /**
+     * @var string
+     */
+    public $threshold;
+    protected $_name = [
+        'operator'  => 'Operator',
+        'option'    => 'Option',
+        'threshold' => 'Threshold',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->operator) {
+            $res['Operator'] = $this->operator;
+        }
+        if (null !== $this->option) {
+            $res['Option'] = $this->option;
+        }
+        if (null !== $this->threshold) {
+            $res['Threshold'] = $this->threshold;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return diagnosis
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Operator'])) {
+            $model->operator = $map['Operator'];
+        }
+        if (isset($map['Option'])) {
+            $model->option = $map['Option'];
+        }
+        if (isset($map['Threshold'])) {
+            $model->threshold = $map['Threshold'];
+        }
+
+        return $model;
+    }
+}