+++ ehpc-20240730/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-12-09 Version: 2.0.3
+- Update API ListJobs: update param JobFilter.
+- Update API ListJobs: update response param.
+
+
 2024-12-05 Version: 2.0.2
 - Update API GetCluster: update response param.
 - Update API UpdateCluster: add param MonitorSpec.